+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_file_system_client.hpp
@@ -215,6 +215,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
           m_pipeline(std::move(pipeline))
     {
     }
+    friend class DataLakeLeaseClient;
     friend class DataLakeServiceClient;
   };
 }}}} // namespace Azure::Storage::Files::DataLake